Added StudyHelper Windows downloads
RandomGameChooser now does not datestamp its updates as it does not have
an automatic updater (no need).  I also added StudyHelper and a zip of
some sample exercises for it.  I also added a datestamp for StudyHelper
so that I can make a launcher for it, as I seek to update it a lot.